Remove width and height (#41)
Having the width and height set here is causing these icons to act
strangely in when set as a background property. I checked this for ie11
and chrome and the problem is present in both browsers. It appears as
though the width and height are not changed by the css so stay big.
Sketch adds these properties.

Use the following css (and an svg with a width and height) to replicate:

```css
.icon-test {
	width: 40px;
	height: 40px;
	border: dotted 1px black;
	display: inline-block;
	background-repeat: no-repeat;
	// Doesn't work in ie8/7/6 in these cases the icon will fill the available space
	background-size: contain;
	background-position: 50%;
	background-color: transparent;
	vertical-align: baseline;

	bottom: 10px;
	left: 10px;
	padding: 10px 16px 10px 70px;
	background-position: left;
	background-image: url(path/to/your/icon.svg)
}
```